streamlit/requirements.txt
streamlit==1.29.0
fastapi==0.95.0
uvicorn==0.21.1
requests==2.31.0
python-jose[cryptography]==3.3.0
httpx==0.24.0